Move submission block assembly code to base libblkmaker
